--- a/src/test/java/org/apache/sysds/performance/compression/APerfTest.java
+++ b/src/test/java/org/apache/sysds/performance/compression/APerfTest.java
@@ -23,6 +23,7 @@ import java.util.ArrayList;
 
 import org.apache.sysds.performance.TimingUtils;
 import org.apache.sysds.performance.TimingUtils.F;
+import org.apache.sysds.performance.TimingUtils.StatsType;
 import org.apache.sysds.performance.generators.IGenerate;
 
 public abstract class APerfTest<T, G> {
@@ -39,19 +40,27 @@ public abstract class APerfTest<T, G> {
 	/** Warmup iterations */
 	protected final int W;
 
+	/** The type of statistics to use */
+	protected final StatsType st;
+
 	protected APerfTest(int N, IGenerate<G> gen) {
-		ret = new ArrayList<>(N);
-		this.gen = gen;
-		this.N = N;
-		this.W = 10;
+		this(N, 10, gen, StatsType.MEAN_STD);
 	}
 
+	protected APerfTest(int N, IGenerate<G> gen, StatsType st) {
+		this(N, 10, gen, st);
+	}
 
 	protected APerfTest(int N, int W, IGenerate<G> gen) {
+		this(N, W, gen, StatsType.MEAN_STD);
+	}
+
+	protected APerfTest(int N, int W, IGenerate<G> gen, StatsType st) {
 		ret = new ArrayList<>(N);
 		this.gen = gen;
 		this.N = N;
-		this.W = 10;
+		this.W = W;
+		this.st = st;
 	}
 
 	protected void execute(F f, String name) throws InterruptedException {
@@ -70,7 +79,7 @@ public abstract class APerfTest<T, G> {
 		ret.clear();
 		double[] times = TimingUtils.time(f, c, b, N, gen);
 		String retS = makeResString(times);
-		System.out.println(String.format("%35s, %s, %10s", name, TimingUtils.stats(times), retS));
+		System.out.println(String.format("%35s, %s, %10s", name, TimingUtils.stats(times, st), retS));
 	}
 
 	protected void warmup(F f, int n) throws InterruptedException {
@@ -92,7 +101,7 @@ public abstract class APerfTest<T, G> {
 		ret.clear();
 		double[] times = TimingUtils.time(f, c, b, N, gen);
 		String retS = makeResString(times);
-		System.out.println(String.format("%35s, %s, %10s", name, TimingUtils.stats(times), retS));
+		System.out.println(String.format("%35s, %s, %10s", name, TimingUtils.stats(times, st), retS));
 	}
 
 	protected abstract String makeResString();
@@ -106,7 +115,7 @@ public abstract class APerfTest<T, G> {
 		StringBuilder sb = new StringBuilder();
 		sb.append(String.format("%20s ", this.getClass().getSimpleName()));
 		sb.append(" Repetitions: ").append(N).append("\n");
-		sb.append(String.format("%20s ","Generator:"));
+		sb.append(String.format("%20s ", "Generator:"));
 		sb.append(gen);
 		sb.append("\n");
 		return sb.toString();